Restart network on CentOS6-Teeth for IPv6
On first boot, on CentOS6-Teeth IPv6 does not work until you do a
restrt of the network service. Prior to this, th eIPv6 address is set,
and routes are in place, but the network is unreachable, except, oddly,
the gateway.

After a network reset, everything works.
This is a work-around that just issues a network restart from the
cloud-init per-instance scripts. Ultimately, this may be kernel-related,
and we should check if this is still needed if we unpin the CentOS 6 kernel.

JIRA:IMG-240